requirements.txt
uuid
redis
pymysql
aiofiles
tornado
requests
timer
torch
numpy
dbutils
python-dotenv
